+++ src/main/java/org/sparta/mytaek1/Mytaek1Application.java
@@ -3,9 +3,11 @@ package org.sparta.mytaek1;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableJpaAuditing
+@EnableScheduling
 public class Mytaek1Application {
 
     public static void main(String[] args) {
