+++ src/main/java/org/sparta/mytaek1/domain/broadcast/entity/Broadcast.java
@@ -18,9 +18,7 @@ public class Broadcast extends Auditable {
     private Long broadcastId;
 
     private String broadcastTitle;
-
     private String broadcastDescription;
-
     private boolean onAir;
 
     @ManyToOne(fetch = FetchType.LAZY)
