+++ src/main/java/org/sparta/mytaek1/domain/order/controller/OrderController.java
@@ -4,8 +4,6 @@ import lombok.RequiredArgsConstructor;
 import org.sparta.mytaek1.domain.order.dto.OrderRequestDto;
 import org.sparta.mytaek1.domain.order.dto.OrderResponseDto;
 import org.sparta.mytaek1.domain.order.service.OrderService;
-import org.sparta.mytaek1.domain.product.entity.Product;
-import org.sparta.mytaek1.domain.product.service.ProductService;
 import org.sparta.mytaek1.domain.stock.entity.Stock;
 import org.sparta.mytaek1.domain.stock.service.StockService;
 import org.sparta.mytaek1.domain.user.entity.User;
@@ -24,7 +22,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class OrderController {
 
     private final OrderService orderService;
-    private final ProductService productService;
     private final StockService stockService;
 
     @PostMapping("/products/{productId}/orders")
