+++ src/main/java/org/sparta/mytaek1/domain/order/entity/Orders.java
@@ -19,7 +19,6 @@ public class Orders extends Auditable {
     private Long orderId;
 
     private Integer quantity;
-
     private Integer totalPrice;
 
     @ManyToOne(fetch = FetchType.LAZY)
