+++ src/main/java/org/sparta/mytaek1/domain/order/repository/OrderRepository.java
@@ -1,19 +1,15 @@
 package org.sparta.mytaek1.domain.order.repository;
 
 import org.sparta.mytaek1.domain.order.entity.Orders;
-import org.sparta.mytaek1.domain.product.entity.Product;
-import org.sparta.mytaek1.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Orders, Long> {
-    Orders findByProductAndUser(Product product, User user);
-
-    Optional<Orders> findByProduct(Product product);
-
     List<Orders> findAllByProductProductId(Long productId);
+
+    List<Orders> findByPaymentStatusAndCreatedAtBefore(boolean status, LocalDateTime tenMinutesAgo);
 }
