+++ src/main/java/org/sparta/mytaek1/domain/order/service/OrderService.java
@@ -2,25 +2,23 @@ package org.sparta.mytaek1.domain.order.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.redisson.api.RedissonClient;
 import org.sparta.mytaek1.domain.order.dto.OrderRequestDto;
 import org.sparta.mytaek1.domain.order.dto.OrderResponseDto;
 import org.sparta.mytaek1.domain.order.entity.Orders;
 import org.sparta.mytaek1.domain.order.repository.OrderRepository;
 import org.sparta.mytaek1.domain.product.entity.Product;
-import org.sparta.mytaek1.domain.product.repository.ProductRepository;
 import org.sparta.mytaek1.domain.product.service.ProductService;
 import org.sparta.mytaek1.domain.stock.entity.Stock;
-import org.sparta.mytaek1.domain.stock.repository.StockRepository;
 import org.sparta.mytaek1.domain.stock.service.StockService;
 import org.sparta.mytaek1.domain.user.entity.User;
-import org.sparta.mytaek1.global.config.RedisConfig;
 import org.sparta.mytaek1.global.message.ErrorMessage;
 import org.sparta.mytaek1.global.redis.DistributedLock;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -31,67 +29,36 @@ public class OrderService {
     private final ProductService productService;
     private final StockService stockService;
 
-//    public OrderResponseDto createOrder(Long productId, OrderRequestDto orderRequestDto, User user) {
-//        RLock fairLock = redissonClient.getFairLock("pro:" + productId);
-//        Product product = productRepository.findById(productId).orElseThrow();
-//        Stock stock = stockRepository.findByProductProductId(productId).orElseThrow();
-//        Orders order = new Orders(orderRequestDto, product, user, false);
-//        System.out.println("하이요");
-//        try{
-//            if(fairLock.tryLock(1000000, TimeUnit.SECONDS)){
-//                System.out.println("겟락");
-//                stock.updateStock(orderRequestDto.getQuantity());
-//                orderRepository.save(order);
-//                Thread.sleep(100);
-//            }
-//        } catch (InterruptedException e) {
-//            System.out.println(e);
-//            throw new RuntimeException(e);
-//        } finally {
-//            if (fairLock.isHeldByCurrentThread()) {
-//                System.out.println("드랍락");
-//                fairLock.unlock();
-//            }
-//
-//        }
-//        return new OrderResponseDto(order);
-//    }
-
-//    @DistributedLock(key = "#lockName")
-//    public OrderResponseDto createOrder(String lockName, Long productId, OrderRequestDto orderRequestDto, User user) {
-//        Product product = findProduct(productId);
-//        Stock stock = findStock(productId);
-//        Orders order = new Orders(orderRequestDto, product, user, false);
-//        stock.updateStock(orderRequestDto.getQuantity());
-//        orderRepository.save(order);
-//        return new OrderResponseDto(order);
-//    }
-
     @DistributedLock(key = "#lockName")
     public OrderResponseDto createOrder(long lockName, Long productId, OrderRequestDto orderRequestDto, User user) {
-        log.info(lockName + "가 락 획득 성공"+ LocalDateTime.now());
         Product product = productService.findProduct(productId);
         Stock stock = stockService.findStockById(productId);
+
         if (stock.getProductStock() < orderRequestDto.getQuantity()) {
-            throw new IllegalArgumentException("잔여 수량이 부족합니다.");
+            throw new IllegalArgumentException(ErrorMessage.NOT_EXIST_STOCK_ERROR_MESSAGE.getErrorMessage());
         }
+
         Orders order = new Orders(orderRequestDto, product, user, false);
         stock.updateStock(orderRequestDto.getQuantity());
         orderRepository.save(order);
+
         return new OrderResponseDto(order);
     }
 
+    @Scheduled(fixedRate = 10000)
     @Transactional
-    public OrderResponseDto createOrderBeforeLock(Long productId, OrderRequestDto orderRequestDto, User user) {
-        Product product = productService.findProduct(productId);
-        Stock stock = stockService.findStockById(productId);
-        if (stock.getProductStock() < orderRequestDto.getQuantity()) {
-            throw new IllegalArgumentException("잔여 수량이 부족합니다.");
+    public void deleteOrder() {
+        LocalDateTime now = LocalDateTime.now();
+        LocalDateTime tenMinutesAgo = now.minusMinutes(10);
+        List<Orders> ordersToDelete = orderRepository.findByPaymentStatusAndCreatedAtBefore(false, tenMinutesAgo);
+
+        for (Orders order : ordersToDelete) {
+            if (order.getCreatedAt().plusMinutes(10).isBefore(now)) {
+                Stock stock = stockService.findStockByProduct(order.getProduct().getProductId());
+                stock.cancelStock(order.getQuantity());
+                orderRepository.delete(order);
+            }
         }
-        Orders order = new Orders(orderRequestDto, product, user, false);
-        stock.updateStock(orderRequestDto.getQuantity());
-        orderRepository.save(order);
-        return new OrderResponseDto(order);
     }
 
     @Transactional(readOnly = true)
@@ -101,9 +68,6 @@ public class OrderService {
     }
 
     public Orders findOrderById(Long orderId) {
-        Orders order = orderRepository.findById(orderId).orElseThrow(()-> new NullPointerException(ErrorMessage.NOT_EXIST_ORDER_ERROR_MESSAGE.getErrorMessage()));
-        return order;
+        return orderRepository.findById(orderId).orElseThrow(()-> new NullPointerException(ErrorMessage.NOT_EXIST_ORDER_ERROR_MESSAGE.getErrorMessage()));
     }
-
-
 }
