+++ src/main/java/org/sparta/mytaek1/domain/payment/controller/PaymentController.java
@@ -6,9 +6,11 @@ import com.siot.IamportRestClient.request.BillingCustomerData;
 import com.siot.IamportRestClient.response.BillingCustomer;
 import com.siot.IamportRestClient.response.IamportResponse;
 import com.siot.IamportRestClient.response.Payment;
+import jakarta.annotation.PostConstruct;
 import org.sparta.mytaek1.domain.order.dto.OrderResponseDto;
 import org.sparta.mytaek1.domain.payment.dto.PaymentRequestDto;
 import org.sparta.mytaek1.domain.payment.service.PaymentService;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,13 +19,21 @@ import java.io.IOException;
 @RestController
 public class PaymentController {
 
-    private final IamportClient iamportClient;
     private final PaymentService paymentService;
+    @Value("${iamport.api.key}")
+    private String apiKey;
+
+    @Value("${iamport.api.secret.key}")
+    private String apiSecretKey;
+    private IamportClient iamportClient;
 
     public PaymentController(PaymentService paymentService) {
         this.paymentService = paymentService;
-        this.iamportClient = new IamportClient("8474815404642414",
-                "iKb4UTArrJKpuSXzivlsLheOgW4NwzVk2pfgxWqKkS4W428lwpwdTp5Xo7cozxjnaw5vJAmIHuvkN6Do");
+    }
+
+    @PostConstruct
+    public void init() {
+        this.iamportClient = new IamportClient(apiKey, apiSecretKey);
     }
 
     @ResponseBody
