--- src/main/java/org/sparta/mytaek1/domain/product/dto/UpdateProductStockRequestDto.java
@@ -1,4 +1,0 @@
-package org.sparta.mytaek1.domain.product.dto;
-
-public class UpdateProductStockRequestDto {
-}
