+++ src/main/java/org/sparta/mytaek1/domain/product/repository/ProductRepository.java
@@ -4,10 +4,7 @@ import org.sparta.mytaek1.domain.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
-    Optional<Product> findByProductName(String productName);
 }
