+++ src/main/java/org/sparta/mytaek1/domain/product/service/ProductService.java
@@ -3,12 +3,9 @@ package org.sparta.mytaek1.domain.product.service;
 import lombok.RequiredArgsConstructor;
 import org.sparta.mytaek1.domain.broadcast.dto.BroadcastRequestDto;
 import org.sparta.mytaek1.domain.product.dto.ProductFindResponseDto;
-import org.sparta.mytaek1.domain.product.dto.UpdateProductStockRequestDto;
-import org.sparta.mytaek1.domain.product.dto.UpdateProductStockResponseDto;
 import org.sparta.mytaek1.domain.product.entity.Product;
 import org.sparta.mytaek1.domain.product.repository.ProductRepository;
 import org.sparta.mytaek1.domain.stock.entity.Stock;
-import org.sparta.mytaek1.domain.stock.repository.StockRepository;
 import org.sparta.mytaek1.domain.stock.service.StockService;
 import org.sparta.mytaek1.global.message.ErrorMessage;
 import org.springframework.stereotype.Service;
@@ -17,6 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class ProductService {
+
     private final ProductRepository productRepository;
     private final StockService stockService;
 
@@ -25,10 +23,6 @@ public class ProductService {
         Stock stock = stockService.findStockById(productId);
         Product product = findProduct(productId);
         return new ProductFindResponseDto(product, stock);
-    }
-
-    public UpdateProductStockResponseDto updateProductStock(UpdateProductStockRequestDto requestDto) {
-        return null;
     }
 
     @Transactional
