+++ src/main/java/org/sparta/mytaek1/domain/user/jwt/JwtUtil.java
@@ -31,7 +31,8 @@ public class JwtUtil {
     private static final String AUTHORIZATION_KEY = "auth";
     private static final String BEARER_PREFIX = "Bearer ";
     private static final String JWT_LOG_HEAD = "JWT 관련 로그";
-    private final long TOKEN_TIME = 60 * 60 * 1000L; // 60분
+    private final long TOKEN_TIME = 60 * 60 * 1000L;
+
     @Value("${jwt.secret.key}")
     private String secretKey;
     private Key key;
