--- src/main/java/org/sparta/mytaek1/global/config/RedisConfig.java
@@ -1,22 +1,0 @@
-package org.sparta.mytaek1.global.config;
-
-import lombok.Getter;
-import org.redisson.Redisson;
-import org.redisson.api.RedissonClient;
-import org.redisson.config.Config;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-public class RedisConfig {
-    @Getter
-    private static final RedissonClient redissonClient = createRedissonClient();
-
-    private static RedissonClient createRedissonClient() {
-        Config config = new Config();
-        config.useSingleServer()
-                .setAddress("redis://127.0.0.1:6379");
-        return Redisson.create(config);
-    }
-
-}
