+++ src/main/java/org/sparta/mytaek1/global/config/RedissonConfig.java
@@ -9,6 +9,7 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class RedissonConfig {
+
     @Value("${spring.data.redis.host}")
     private String redisHost;
