+++ src/main/java/org/sparta/mytaek1/global/redis/AopForTransaction.java
@@ -5,9 +5,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-/**
- * AOP에서 트랜잭션 분리를 위한 클래스
- */
 @Component
 public class AopForTransaction {
