+++ src/main/java/org/sparta/mytaek1/global/redis/CustomSpringELParser.java
@@ -1,16 +1,12 @@
 package org.sparta.mytaek1.global.redis;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.expression.ExpressionParser;
 import org.springframework.expression.spel.standard.SpelExpressionParser;
 import org.springframework.expression.spel.support.StandardEvaluationContext;
 
-/**
- * Spring Expression Language Parser
- */
+@RequiredArgsConstructor
 public class CustomSpringELParser {
-    private CustomSpringELParser() {
-    }
-
     public static Object getDynamicValue(String[] parameterNames, Object[] args, String key) {
         ExpressionParser parser = new SpelExpressionParser();
         StandardEvaluationContext context = new StandardEvaluationContext();
