+++ src/main/java/org/sparta/mytaek1/global/security/UserRoleEnum.java
@@ -1,16 +1,16 @@
 package org.sparta.mytaek1.global.security;
 
+import lombok.Getter;
+
+@Getter
 public enum UserRoleEnum {
+
     USER(Authority.USER);
 
     private final String authority;
 
     UserRoleEnum(String authority) {
         this.authority = authority;
-    }
-
-    public String getAuthority() {
-        return this.authority;
     }
 
     public static class Authority {
