+++ src/main/resources/templates/broadcastForm.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Broadcast Form</title>
+    <title>MyTaek1 - Broadcast Form</title>
     <script src="https://code.jquery.com/jquery-3.6.4.min.js"></script>
     <script type="text/javascript" th:src="@{/js/broadcastForm_js.js}">
 </script>
