+++ src/main/resources/templates/broadcastList.html
@@ -2,33 +2,32 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
-    <title>Broadcast List</title>
+    <title>MyTaek1 - Broadcast List</title>
 </head>
 <body>
     <nav class="navbar"></nav>
     <h2>Broadcast List</h2>
     <table border="1">
         <thead>
-        <tr>
-            <th>Title</th>
-            <th>Description</th>
-            <th>User Name</th>
-            <th>Product Name</th>
-        </tr>
+            <tr>
+                <th>Title</th>
+                <th>Description</th>
+                <th>User Name</th>
+                <th>Product Name</th>
+            </tr>
         </thead>
         <tbody>
-
-        <tr th:each="broadcastResponseDto : ${broadcastResponseDtoList}">
-            <td th:text="${broadcastResponseDto.broadCastTitle}"></td>
-            <td th:text="${broadcastResponseDto.broadCastDescription}"></td>
-            <td th:text="${broadcastResponseDto.userName}"></td>
-            <td th:text="${broadcastResponseDto.productName}"></td>
-            <td>
-                <button class="view-details-button" th:attr="data-broadcast-id=${broadcastResponseDto.broadcastId}" style="cursor: pointer;">
-                    View Details
-                </button>
-            </td>
-        </tr>
+            <tr th:each="broadcastResponseDto : ${broadcastResponseDtoList}">
+                <td th:text="${broadcastResponseDto.broadCastTitle}"></td>
+                <td th:text="${broadcastResponseDto.broadCastDescription}"></td>
+                <td th:text="${broadcastResponseDto.userName}"></td>
+                <td th:text="${broadcastResponseDto.productName}"></td>
+                <td>
+                    <button class="view-details-button" th:attr="data-broadcast-id=${broadcastResponseDto.broadcastId}" style="cursor: pointer;">
+                        View Details
+                    </button>
+                </td>
+            </tr>
         </tbody>
     </table>
     <button onclick="window.location.href='/broadcasts/start'">방송 시작하기</button>
