+++ src/main/resources/templates/myPage.html
@@ -9,13 +9,10 @@
   <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
   <title>MyTaek1 - MyPage</title>
   <script th:inline="javascript">
-    /*<![CDATA[*/
-    /* Reload the page only once */
     if (!sessionStorage.getItem('reloaded')) {
       location.reload();
       sessionStorage.setItem('reloaded', 'true');
     }
-    /*]]>*/
   </script>
 </head>
 <body>
@@ -29,7 +26,6 @@
     <p>Your Phone: <span th:text="${userPhone}"></span></p>
     <p>Your Address: <span th:text="${userAddress}"></span></p>
     <p>Your Postcode: <span th:text="${postcode}"></span></p>
-
     <div class="flex-box">
       <div class="broadcast_container">
         <h2 style="margin: 40px 0 20px 0;">Your Broadcasts</h2>
