--- src/test/java/org/sparta/mytaek1/domain/product/ProduceServiceTest.java
@@ -1,4 +1,0 @@
-package org.sparta.mytaek1.domain.product;
-
-public class ProduceServiceTest {
-}
